+++ app/components/NavBar.js
@@ -1,17 +1,25 @@
 import Head from 'next/head'
 import styles from '../styles/Navbar.module.scss'
 import { AuthContext } from '../contexts/AuthContext'
-import SignIn from './SignInModal'
 import { useContext } from 'react'
+import { useRouter } from 'next/router'
 
 const Navbar = () => {
-  const { setSignInModalOpen } = useContext(AuthContext)
-
+  const { user, setSignInModalOpen } = useContext(AuthContext)
+  const router = useRouter()
+  const handleOpen = () => {
+    if (user) {
+      router.push('/mypage')
+    } else {
+      setSignInModalOpen(true)
+    }
+  }
   return (
     <div className={styles.container}>
       <div>hello nav bar</div>
-      <button onClick={() => setSignInModalOpen(true)}>open</button>
-      <SignIn />
+      <button onClick={handleOpen} type="button">
+        open
+      </button>
     </div>
   )
 }
