+++ app/components/PageTransition.js
@@ -20,7 +20,7 @@ const Transition = ({ children, ...props }) => {
       right: 0,
       bottom: 0,
       left: 0,
-      opacity: 0
+      opacity: 1
     }
   })
